Sprite compatibility with thumbyGraphics
Grayscale Sprites backwards compatible enough to be passed
to thumbyGraphics methods to show the black and white equivalent.

Removal of the ability to pass a tuple containing both bitmap
layers to the mask parameter of blitWithMask. This didn't make
much sense on its own and was previously used to
enable compatibility with supporting black and white Sprites,
but these changes make this unneccessary.