Fix the config reload crash that affects a small percentage of Thumbys.
These thumbys would work correctly with grayscale until the calibration
config is loaded on a second run. Grayscale would then never load correctly
until the config file is deleted and it is calibrated again.

This emission of a short inaudible burst of audio seems to fix the issue.
I have no idea why. But it does, for all affected Thumbys:
https://discord.com/channels/898292107289190461/898292174410612787/1299077610025783317
https://discord.com/channels/898292107289190461/898292174410612787/1299091476239614155
https://discord.com/channels/898292107289190461/996871099499425873/1256231698098688052
Maybe it's a timing issue on loading config files.
Maybe it is that the flash isn't initialised properly at that point.
Maybe the audio burst, gives the flash a little nudge to shake it awake.
I don't know, but until I have access to an affected Thumby, or until
someone wants to golf this, this will have to do.

Thanks to PF (@htmlgames from the Discord channel) for this magic fix!